+++ TombEngine/Scripting/Internal/TEN/Effects/EffectsFunctions.cpp
@@ -10,7 +10,6 @@
 #include "Game/effects/Electricity.h"
 #include "Game/effects/explosion.h"
 #include "Game/effects/spark.h"
-#include "Game/effects/Streamer.h"
 #include "Game/effects/tomb4fx.h"
 #include "Game/effects/weather.h"
 #include "Game/Setup.h"
@@ -21,9 +20,7 @@
 #include "Scripting/Internal/ScriptUtil.h"
 #include "Scripting/Internal/TEN/Effects/BlendIDs.h"
 #include "Scripting/Internal/TEN/Effects/EffectIDs.h"
-#include "Scripting/Internal/TEN/Effects/FeatherModes.h"
-#include "Scripting/Internal/TEN/Objects/Moveable/MoveableObject.h"
-#include "Scripting/Internal/TEN/Objects/ObjectsHandler.h"
+#include "Scripting/Internal/TEN/Effects/ParticleAnimTypes.h"
 #include "Scripting/Internal/TEN/Types/Color/Color.h"
 #include "Scripting/Internal/TEN/Types/Vec3/Vec3.h"
 #include "Scripting/Internal/TEN/Types/Vec2/Vec2.h"
@@ -41,7 +38,6 @@ using namespace TEN::Effects::Electricity;
 using namespace TEN::Effects::Environment;
 using namespace TEN::Effects::Explosion;
 using namespace TEN::Effects::Spark;
-using namespace TEN::Effects::Streamer;
 using namespace TEN::Math;
 using namespace TEN::Scripting::Types;
 
@@ -215,6 +211,156 @@ namespace TEN::Scripting::Effects
 		if (TestEnvironment(RoomEnvFlags::ENV_FLAG_WIND, part.roomNumber))
 			part.flags |= SP_WIND;
 	}
+
+	/// Emit a particle with extensive configuration options including animations.
+	// @function EmitAdvancedParticle
+	// @tparam ParticleData ParticleData The table holding all the particle data.
+	// @usage
+	// local particle = {
+	// position = GetMoveableByName("camera_target_6") :GetPosition(), 
+	// velocity = Vec3(0, 0, 10),
+	// spriteSeqID = TEN.Objects.ObjID.CUSTOM_BAR_GRAPHIC,
+	// spriteID = 0,
+	// lifetime = 10,
+	// maxYVelocity = 0,
+	// gravity = 0,
+	// friction = 10,
+	// startRotation = 0,
+	// rotationSpeed = 0,
+	// startSize = 80,
+	// endSize = 80,
+	// startColor = TEN.Color(128, 128, 128),
+	// endColor = TEN.Color(128, 128, 128),
+	// blendMode = TEN.Effects.BlendID.ADDITIVE,
+	// wind = false,
+	// damage = true,
+	// poison = false,
+	// burn = false,
+	// damageHit = 80,
+	// sound = 197,
+	// light = true,
+	// lightRadius = 6, 
+	// lightFlicker = 5, 
+	// animated = true,
+	// frameRate = .25,
+	// animationType = TEN.Effects.ParticleAnimationType.LOOP,
+	// }
+	// EmitAdvancedParticle(particle)
+	static void EmitAdvancedParticle(sol::table ParticleData)
+	{
+		constexpr auto DEFAULT_START_SIZE = 10.0f;
+		constexpr auto DEFAULT_LIFE = 2.0f;
+		constexpr auto SECS_PER_FRAME = 1.0f / (float)FPS;
+
+		auto convertedSpriteSeqID = ParticleData.get_or("spriteSeqID", ID_DEFAULT_SPRITES);
+		if (!CheckIfSlotExists(convertedSpriteSeqID, "EmitParticle() script function."))
+			return;
+
+		auto& part = *GetFreeParticle();
+
+		part.on = true;
+		part.SpriteSeqID = convertedSpriteSeqID;
+		part.SpriteID = ParticleData.get_or("spriteID", 0);
+
+		auto bMode = ParticleData.get_or("blendMode", BlendMode::AlphaBlend);
+		part.blendMode = bMode;
+
+		Vec3 pos = ParticleData["position"];
+		part.x = pos.x;
+		part.y = pos.y;
+		part.z = pos.z;
+		part.roomNumber = FindRoomNumber(Vector3i(pos.x, pos.y, pos.z));
+
+		Vec3 vel = ParticleData["velocity"];
+		part.xVel = short(vel.x * 32);
+		part.yVel = short(vel.y * 32);
+		part.zVel = short(vel.z * 32);
+
+		float rotAdd = ParticleData.get_or("rotationSpeed", 0.0f);
+		float rotAng = ParticleData.get_or("startRotation", TO_DEGREES(Random::GenerateAngle()));
+		part.rotAng = ANGLE(rotAng) >> 4;
+		part.rotAdd = ANGLE(rotAdd) >> 4;
+
+		part.sSize =
+			part.size = ParticleData.get_or("startSize", DEFAULT_START_SIZE);
+		part.dSize = ParticleData.get_or("endSize", 0.0f);
+		part.scalar = 2;
+
+		part.gravity = (short)std::clamp((float) ParticleData.get_or("gravity", 0.0f), (float)SHRT_MIN, (float)SHRT_MAX);
+		part.friction = ParticleData.get_or("friction", 0);
+		part.maxYvel = ParticleData.get_or("maxYVelocity", 0);
+
+		auto convertedStartColor = ParticleData.get_or("startColor", ScriptColor(255, 255, 255));
+		part.sR = convertedStartColor.GetR();
+		part.sG = convertedStartColor.GetG();
+		part.sB = convertedStartColor.GetB();
+
+		auto convertedEndColor = ParticleData.get_or("endColor", ScriptColor(255, 255, 255));
+		part.dR = convertedEndColor.GetR();
+		part.dG = convertedEndColor.GetG();
+		part.dB = convertedEndColor.GetB();
+
+		float convertedLife = std::max(0.1f, (float) ParticleData.get_or("lifetime", DEFAULT_LIFE));
+		part.life =
+			part.sLife = (int)round(convertedLife / SECS_PER_FRAME);
+		part.colFadeSpeed = part.life / 2;
+		part.fadeToBlack = part.life / 3;
+
+		part.flags = SP_SCALE | SP_ROTATE | SP_DEF | SP_EXPDEF;
+
+		part.damage = ParticleData.get_or("damageHit", 2);
+
+		bool convertedApplyPoison = ParticleData.get_or("poison", false);
+		if (convertedApplyPoison)
+			part.flags |= SP_POISON;
+
+		bool convertedApplyDamage = ParticleData.get_or("damage", false);
+		if (convertedApplyDamage)
+			part.flags |= SP_DAMAGE;
+
+		bool convertedApplyBurn = ParticleData.get_or("burn", false);
+		if (convertedApplyBurn)
+			part.flags |= SP_FIRE;
+
+		int convertedApplySound = ParticleData.get_or("sound", -1);
+		if (convertedApplySound > 0)
+		{
+			part.flags |= SP_SOUND;
+			part.sound = convertedApplySound;
+		}
+		bool convertedApplyLight = ParticleData.get_or("light", false);
+		if (convertedApplyLight)
+		{
+			part.flags |= SP_LIGHT;
+			int lightRadius = ParticleData.get_or("lightRadius", 0);
+			part.lightRadius = lightRadius * BLOCK(0.25f);
+			int flicker = ParticleData.get_or("lightFlicker", 0);
+			
+			if (flicker > 0)
+			{
+				part.lightFlicker = ParticleData.get_or("lightFlicker", 0);
+				part.lightFlickerS = ParticleData.get_or("lightFlicker", 0);
+			}
+		}
+		bool animatedSpr = ParticleData.get_or("animated", false);
+		if (animatedSpr)
+		{
+			ParticleAnimType applyAnim = ParticleData.get_or("animationType", ParticleAnimType::Loop);
+			float applyFramerate = ParticleData.get_or("frameRate", 1.0f);
+			part.flags |= SP_ANIMATED;
+			part.framerate = applyFramerate;
+			part.animationType = ParticleAnimType(std::clamp(int(applyAnim), int(ParticleAnimType::None), int(ParticleAnimType::LifetimeSpread)));
+
+		}
+
+		bool convertedApplyWind = ParticleData.get_or("wind", false);
+		if (convertedApplyWind)
+		{
+			if (TestEnvironment(RoomEnvFlags::ENV_FLAG_WIND, part.roomNumber))
+				part.flags |= SP_WIND;
+		}
+
+	}
 	
 /***Emit a shockwave, similar to that seen when a harpy projectile hits something.
 	@function EmitShockwave
@@ -304,11 +450,11 @@ namespace TEN::Scripting::Effects
 		TriggerBlood(pos.x, pos.y, pos.z, -1, ValueOr<int>(count, 1));
 	}
 
-	/// Emit air bubble in a water room.
-	// @function EmitAirBubble
-	// @tparam Vec3 pos World position where the effect will be spawned. Must be in a water room.
-	// @tparam[opt] float size Sprite size. __Default: 32__
-	// @tparam[opt] float amp Oscillation amplitude. __Default: 32__
+/// Emit air bubble in a water room.
+// @function EmitAirBubble
+// @tparam Vec3 pos World position where the effect will be spawned. Must be in a water room.
+// @tparam[opt] float size Sprite size. __Default: 32__
+// @tparam[opt] float amp Oscillation amplitude. __Default: 32__
 	static void EmitAirBubble(const Vec3& pos, TypeOrNil<float> size, TypeOrNil<float> amp)
 	{
 		constexpr auto DEFAULT_SIZE = 128.0f;
@@ -328,49 +474,6 @@ namespace TEN::Scripting::Effects
 	static void EmitFire(const Vec3& pos, TypeOrNil<float> size)
 	{
 		AddFire(pos.x, pos.y, pos.z, FindRoomNumber(Vector3i(pos.x, pos.y, pos.z)), ValueOr<float>(size, 1));
-	}
-
-	/// Emit an extending streamer effect.
-	// @function EmitStreamer
-	// @tparam Moveable mov Moveable object with which to associate the effect.
-	// @tparam int[opt] tag Numeric tag with which to associate the effect on the moveable. __default: 0__
-	// @tparam Vec3 pos World position.
-	// @tparam Vec3 dir Direction vector of movement velocity.
-	// @tparam[opt] float rot Start rotation in degrees. __default: 0__
-	// @tparam[opt] Color startColor Color at the start of life. __default: Color(255, 255, 255, 255))__
-	// @tparam[opt] Color endColor Color at the end of life. __default: Color(0, 0, 0, 0))__
-	// @tparam[opt] float width Width in world units. __default: 0__
-	// @tparam[opt] float life Lifetime in seconds. __default: 1__
-	// @tparam[opt] float vel Movement velocity in world units per second. __default: 0__
-	// @tparam[opt] float expRate Width expansion rate in world units per second. __default: 0__
-	// @tparam[opt] float rotRate Rotation rate in degrees per second. __default: 0__
-	// @tparam[opt] Effects.FeatherMode featherMode Edge feathering mode. __default: Effects.FeatherMode.NONE__
-	// @tparam[opt] Effects.BlendID blendID Renderer blend ID. __Effects.BlendID.ALPHA_BLEND__
-	static void EmitStreamer(const Moveable& mov, TypeOrNil<int> tag, const Vec3& pos, const Vec3& dir, TypeOrNil<float> rot, TypeOrNil<ScriptColor> startColor, TypeOrNil<ScriptColor> endColor,
-							 TypeOrNil<float> width, TypeOrNil<float> life, TypeOrNil<float> vel, TypeOrNil<float> expRate, TypeOrNil<float> rotRate,
-							 TypeOrNil<StreamerFeatherMode> featherMode, TypeOrNil<BlendMode> blendID)
-	{
-		int movID = mov.GetIndex();
-		int convertedTag = ValueOr<int>(tag, 0);
-		auto convertedPos = pos.ToVector3();
-		auto convertedDir = dir.ToVector3();
-		auto convertedRot = ANGLE(ValueOr<float>(rot, 0));
-		auto convertedStartColor = ValueOr<ScriptColor>(startColor, ScriptColor(255, 255, 255, 255));
-		auto convertedEndColor = ValueOr<ScriptColor>(endColor, ScriptColor(0, 0, 0, 0));
-
-		auto convertedWidth = ValueOr<float>(width, 0.0f);
-		auto convertedLife = ValueOr<float>(life, 1.0f);
-		auto convertedVel = ValueOr<float>(vel, 0.0f) / (float)FPS;
-		auto convertedExpRate = ValueOr<float>(expRate, 0.0f) / (float)FPS;
-		auto convertedRotRate = ANGLE(ValueOr<float>(rotRate, 0.0f) / (float)FPS);
-
-		auto convertedFeatherMode = ValueOr<StreamerFeatherMode>(featherMode, StreamerFeatherMode::None);
-		auto convertedBlendID = ValueOr<BlendMode>(blendID, BlendMode::AlphaBlend);
-
-		StreamerEffect.Spawn(
-			movID, convertedTag, convertedPos, convertedDir, convertedRot, convertedStartColor, convertedEndColor,
-			convertedWidth, convertedLife, convertedVel, convertedExpRate, convertedRotRate,
-			convertedFeatherMode, convertedBlendID);
 	}
 
 /***Make an explosion. Does not hurt Lara
@@ -396,7 +499,7 @@ namespace TEN::Scripting::Effects
 
 	/// Get the wind vector for the current game frame.
 	// This represents the 3D displacement applied by the engine on things like particles affected by wind.
-	// @function GetWind
+	// @function GetWind()
 	// @treturn Vec3 Wind vector.
 	static Vec3 GetWind()
 	{
@@ -408,26 +511,54 @@ namespace TEN::Scripting::Effects
 		auto tableEffects = sol::table(state->lua_state(), sol::create);
 		parent.set(ScriptReserved_Effects, tableEffects);
 
-		// Getters
-		tableEffects.set_function(ScriptReserved_GetWind, &GetWind);
-
-		// Utilities
 		tableEffects.set_function(ScriptReserved_EmitLightningArc, &EmitLightningArc);
 		tableEffects.set_function(ScriptReserved_EmitParticle, &EmitParticle);
+		tableEffects.set_function(ScriptReserved_EmitAdvancedParticle, &EmitAdvancedParticle);
 		tableEffects.set_function(ScriptReserved_EmitShockwave, &EmitShockwave);
 		tableEffects.set_function(ScriptReserved_EmitLight, &EmitLight);
 		tableEffects.set_function(ScriptReserved_EmitSpotLight, &EmitSpotLight);
 		tableEffects.set_function(ScriptReserved_EmitBlood, &EmitBlood);
 		tableEffects.set_function(ScriptReserved_EmitAirBubble, &EmitAirBubble);
-		tableEffects.set_function(ScriptReserved_EmitFire, &EmitFire);
-		tableEffects.set_function(ScriptReserved_EmitStreamer, &EmitStreamer);
 		tableEffects.set_function(ScriptReserved_MakeExplosion, &MakeExplosion);
+		tableEffects.set_function(ScriptReserved_EmitFire, &EmitFire);
 		tableEffects.set_function(ScriptReserved_MakeEarthquake, &Earthquake);
+		tableEffects.set_function(ScriptReserved_GetWind, &GetWind);
 
-		// Enums
 		auto handler = LuaHandler{ state };
 		handler.MakeReadOnlyTable(tableEffects, ScriptReserved_BlendID, BLEND_IDS);
 		handler.MakeReadOnlyTable(tableEffects, ScriptReserved_EffectID, EFFECT_IDS);
-		handler.MakeReadOnlyTable(tableEffects, ScriptReserved_FeatherMode, FEATHER_MODES);
+		handler.MakeReadOnlyTable(tableEffects, ScriptReserved_ParticleAnimationType, PARTICLE_ANIM_TYPES);
 	}
 }
+
+/// Structure for EmitAdvancedParticle table.
+// @table ParticleData
+// @tfield Vec3 position World position.
+// @tfield Vec3 velocity Velocity.
+// @tfield[opt] Objects.ObjID spriteSeqID ID of the sprite sequence object. __Default: Objects.ObjID.DEFAULT_SPRITES__
+// @tfield[opt] int spriteID ID of the sprite in the sprite sequence object.__Default: 0__
+// @tfield[opt] float lifetime Lifespan in seconds. __Default: 2__
+// @tfield[opt] float maxYVelocity Specifies ithe maximum Y velocity for the particle. __Default: 0__
+// @tfield[opt] float gravity Specifies if the particle will fall over time. Positive values ascend, negative values descend. Recommended range: [-1000 and 1000]. __Default: 0__
+// @tfield[opt] float friction Specifies the friction with which the particle will slow down over time. __Default: 0__
+// @tfield[opt] float startRotation Rotation at start of life. __Default: random__
+// @tfield[opt] float rotationSpeed Rotational velocity in degrees. __Default: 0__
+// @tfield[opt] float startSize Size at start of life. __Default: 10__
+// @tfield[opt] float endSize Size at end of life. The particle will linearly shrink or grow toward this size over its lifespan. __Default: 0__
+// @tfield[opt] Color startColor Color at start of life. __Default: Color(255, 255, 255)__
+// @tfield[opt] Color endColor Color to fade toward. This will finish long before the end of the particle's life due to internal math. __Default: Color(255, 255, 255)__
+// @tfield[opt] Effects.BlendID blendMode Render blend mode. __TEN.Effects.BlendID.ALPHA_BLEND__
+// @tfield[opt] bool damage Specify if the particle will harm the player on collision. __Default: false__
+// @tfield[opt] bool poison Specify if the particle will poison the player on collision. __Default: false__
+// @tfield[opt] bool burn Specify if the particle will burn the player on collision. __Default: false__
+// @tfield[opt] bool wind Specify if the particle will be affected by wind in outside rooms. __Default: false__
+// @tfield[opt] int damageHit Specify the damage particle will harm the player on collision. __Default: 2__
+// @tfield[opt] bool light Specify if the particle will be emit a light based on its color. Caution: Recommended only for a single particle. Having too many particles with lights can overflow the light system. __Default: false__
+// @tfield[opt] int lightRadius measured in "clicks" or 256 world units. __Default: 0__
+// @tfield[opt] int lightFlicker The interval at which light should flicker. __Default: 0__
+// @tfield[opt] int sound ID to play. Corresponds to the value in the sound XML file or Tomb Editor's "Sound Infos" window. Looping sounds recommended. Caution: Recommended only for a single particle. Having too many particles with sounds can overflow the sound system. __Default: None__
+// @tfield[opt] bool animated Specify if the particle will be animated. __Default: false__
+// @tfield[opt] Effects.ParticleAnimationType animationType Specify the the type of animation the particle will use. __Default: TEN.Effects.ParticleAnimationType.LOOP__
+// @tfield[opt] float frameRate The framerate with which the particle will be animated. __Default: 1__
+
+